+++ src/platforms/bitflyer/spot/BitFlyerSpotLoginParameters.ts
@@ -20,19 +20,10 @@
  * THE SOFTWARE.
 */
 
-import { Mida, } from "!/src/core/Mida";
-import { logger, } from "#loggers/MidaLogger";
+import { BitFlyerAccountRegion, } from "!/src/platforms/bitflyer/BitFlyerAccountRegion";
 
-// <public-api>
-export * from "!/src/core/Mida";
-
-export * from "!/src/platforms/binance/Binance";
-export * from "!/src/platforms/bitflyer/BitFlyer";
-export * from "!/src/platforms/bybit/Bybit";
-export * from "!/src/platforms/ctrader/CTrader";
-export * from "!/src/platforms/okx/Okx";
-
-export * from "!/src/playground/MidaPlayground";
-// </public-api>
-
-logger.info(`Using Mida ${Mida.version}`);
+export type BitFlyerSpotLoginParameters = {
+    apiKey: string;
+    apiSecret: string;
+    region?: BitFlyerAccountRegion;
+};